Refactored utility "future" to "strings" (and dropped its python2 support)
