+++ kobo/apps/hook/tests/test_parser.py
@@ -8,7 +8,7 @@ import re
 from lxml import etree
 
 from kpi.constants import INSTANCE_FORMAT_TYPE_XML
-from kpi.utils.future import to_str
+from kpi.utils.strings import to_str
 from .hook_test_case import HookTestCase
