+++ kobo/settings/base.py
@@ -165,6 +165,7 @@ CONSTANCE_CONFIG = {
 # Tell django-constance to use a database model instead of Redis
 CONSTANCE_BACKEND = 'constance.backends.database.DatabaseBackend'
 
+
 # Warn developers to use `pytest` instead of `./manage.py test`
 class DoNotUseRunner(object):
     def __init__(self, *args, **kwargs):
@@ -174,9 +175,7 @@ class DoNotUseRunner(object):
 TEST_RUNNER = __name__ + '.DoNotUseRunner'
 
 # used in kpi.models.sitewide_messages
-# markdown does not support imports as unicode.
-# see `python2.7/dist-packages/markitup/fields.py:L14`
-MARKITUP_FILTER = (n(b'markdown.markdown'), {n(b'safe_mode'): False})
+MARKITUP_FILTER = ('markdown.markdown', {'safe_mode': False})
 
 # The backend that handles user authentication must match KoBoCAT's when
 # sharing sessions. ModelBackend does not interfere with object-level
