+++ kpi/tests/api/v2/test_api_asset_snapshots.py
@@ -11,7 +11,7 @@ from rest_framework import status
 from kpi.models.asset import AssetSnapshot
 from kpi.tests.kpi_test_case import KpiTestCase
 from kpi.urls.router_api_v2 import URL_NAMESPACE as ROUTER_URL_NAMESPACE
-from kpi.utils.future import to_str
+from kpi.utils.strings import to_str
 
 
 class TestAssetSnapshotList(KpiTestCase):
