+++ kpi/tests/api/v2/test_api_assets.py
@@ -24,7 +24,7 @@ from kpi.serializers.v2.asset import AssetListSerializer
 from kpi.tests.base_test_case import BaseAssetTestCase, BaseTestCase
 from kpi.tests.kpi_test_case import KpiTestCase
 from kpi.urls.router_api_v2 import URL_NAMESPACE as ROUTER_URL_NAMESPACE
-from kpi.utils.future import to_str, hashable_str
+from kpi.utils.strings import to_str, hashable_str
 
 
 class AssetsListApiTests(BaseAssetTestCase):
