+++ kpi/tests/test_mock_data_conflicting_version_exports.py
@@ -13,7 +13,7 @@ from django.utils.six import text_type
 from kobo.apps.reports import report_data
 from kpi.constants import PERM_VIEW_SUBMISSIONS
 from kpi.models import Asset, ExportTask
-from kpi.utils.future import to_str
+from kpi.utils.strings import to_str
 
 
 class ConflictingVersionsMockDataExports(TestCase):
