+++ kpi/utils/gravatar_url.py
@@ -6,7 +6,7 @@ import hashlib
 
 from django.utils.six.moves.urllib.parse import urlencode
 
-from kpi.utils.future import hashable_str
+from kpi.utils.strings import hashable_str
 
 
 def gravatar_url(email, https=True):
