+++ kpi/utils/mongo_helper.py
@@ -10,7 +10,7 @@ from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext as _
 
 from kpi.constants import NESTED_MONGO_RESERVED_ATTRIBUTES
-from kpi.utils.future import base64_encodestring
+from kpi.utils.strings import base64_encodestring
 
 
 class MongoHelper(object):
