+++ kpi/views/v1/import_task.py
@@ -11,7 +11,7 @@ from rest_framework.reverse import reverse
 from kpi.models import ImportTask
 from kpi.serializers import ImportTaskListSerializer, ImportTaskSerializer
 from kpi.tasks import import_in_background
-from kpi.utils.future import to_str
+from kpi.utils.strings import to_str
 
 
 class ImportTaskViewSet(viewsets.ReadOnlyModelViewSet):
